Revert "inheritance spec: handle a recursive case without polluting the cache with partials."
This commit causes issue https://github.com/spullara/mustache.java/issues/126.
A pull request for this issue was accepted here: https://github.com/spullara/mustache.java/pull/127, but when I cherry-picked that commit, it did not resolve the issue for us. Presumably that pull request one of: 1) somehow only fixes it for 0.9.0, 2) solves a different issue caused by the same commit, or 3) doesn't fix the issue at all.

Since this fixes an issue we already had in mustache 0.8.14 (the version we were previously using), we can live without the fix for now.

This reverts commit c9e64e12d0193ffa5fc2b7fa2c7c0ea05067281d.